+++ src/ts/ir/input.ts
@@ -15,7 +15,7 @@ export const input = (vditor: IVditor, range: Range, ignoreSpace = false) => {
     // 前后可以输入空格
     if (blockElement && !ignoreSpace) {
         if (isHrMD(blockElement.innerHTML) ||
-            isHeadingMD(blockElement.innerHTML, vditor.options.preview.markdown.setext)) {
+            isHeadingMD(blockElement.innerHTML)) {
             return;
         }